using tile endpoint in frontend

+++ backend/app/app/api/tiles_v1/endpoints/tiles.py
@@ -15,25 +15,24 @@ router = APIRouter()
     responses={200: {"content": {"image/png": {}}}},
 )
 def get_tile(
-    zoom: int,
-    xcoord: int,
-    ycoord: int,
+    zoom: int, xcoord: int, ycoord: int,
 ):
     """
     Get tile from cache or render
     """
 
     cached_tile_path = Path("/tiles") / str(zoom) / str(xcoord) / f"{ycoord}.png"
-    try:
-        # try loading file from cache
+    if cached_tile_path.exists():
         with open(cached_tile_path, "rb") as f:
             tile_image = f.read()
-    except FileNotFoundError:
-        # when no cached file found, recreate tile and store to cache
+    else:
         render_task = celery_app.send_task(
             "app.worker.render_tile", args=[zoom, xcoord, ycoord]
         )
-        tile_image = base64.b64decode(render_task.get(timeout=5))
+        # TODO: sometimes the task ran into a timeout
+        # although execution is successful and fast
+        # could not reproduce this error
+        tile_image = base64.b64decode(render_task.get(timeout=10))
         cached_tile_path.parent.mkdir(parents=True, exist_ok=True)
         with open(cached_tile_path, "wb") as f:
             f.write(tile_image)
